+++ Meteoroids/src/test/java/meteoroids/Meteoroids/GameTest.java
@@ -23,7 +23,7 @@ public class GameTest {
     
     @Test
     public void testTimeFactor() {
-        double timeFactor = game.getTimeFactor();
+        float timeFactor = game.getTimeFactor();
         assertEquals("Time factor should be 1.0 as default, but it was "
                 + timeFactor + ".", 1.0, timeFactor, 0.01);
     }
